+++ packages/adhere-ui-backtopanimation/lib/index.js
@@ -1,2 +1,2 @@
-"use strict";var __importDefault=this&&this.__importDefault||function(t){return t&&t.__esModule?t:{default:t}};Object.defineProperty(exports,"__esModule",{value:!0});var backtopanimation_1=__importDefault(require("./backtopanimation"));exports.default=backtopanimation_1.default;
+import BackTopAnimation from"./backtopanimation";export default BackTopAnimation;
 //# sourceMappingURL=index.js.map