+++ packages/adhere-ui-g6/lib/index.js
@@ -1,0 +1,2 @@
+import Template from"./template";export default Template;
+//# sourceMappingURL=index.js.map