+++ packages/adhere-ui-globalindicator/src/index.ts
@@ -1,3 +1,3 @@
-import GlobalIndicator from './globalindicator';
+import './index.less'; import GlobalIndicator from './globalindicator';
 
 export default GlobalIndicator;