+++ packages/adhere-ui-olmap/lib/heatmap.d.ts
@@ -1,3 +1,4 @@
+import { Heatmap as HeatMapLayer } from 'ol/layer';
 import OlMap from './olmap';
 /**
  * HeatMap
@@ -8,7 +9,11 @@ declare class HeatMap extends OlMap {
     static defaultProps: any;
     static propTypes: any;
     private vectorSource;
-    addLayer(): void;
-    getHeatmapLayer(): null;
+    private layer;
+    /**
+     * addLayer - 添加一个热力图图层
+     */
+    addLayer(heatMapLayerConfig?: {}): void;
+    getHeatMapLayer(): HeatMapLayer;
 }
 export default HeatMap;