+++ packages/adhere-ui-olmap/lib/index.js
@@ -1,2 +1,2 @@
-import GeoLayer from"./geolayer";import*as TitleLayer from"./titlelayer";import OLMap from"./olmap";import HeatMap from"./heatmap";import Util from"./util";import AnimationManager from"./animationmanager";export default{AnimationManager:AnimationManager,GeoLayer:GeoLayer,TitleLayer:TitleLayer,OLMap:OLMap,HeatMap:HeatMap,Util:Util};
+import OLMap from"./olmap";import HeatMap from"./heatmap";import GeoLayer from"./geolayer";import*as TitleLayer from"./titlelayer";import Util from"./util";import AnimationManager from"./animationmanager";export default{AnimationManager:AnimationManager,GeoLayer:GeoLayer,TitleLayer:TitleLayer,OLMap:OLMap,HeatMap:HeatMap,Util:Util};
 //# sourceMappingURL=index.js.map