+++ packages/adhere-ui-popup/lib/index.js
@@ -1,2 +1,2 @@
-"use strict";var __importDefault=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(exports,"__esModule",{value:!0});var popup_1=__importDefault(require("./popup"));exports.default=popup_1.default;
+import Popup from"./popup";export default Popup;
 //# sourceMappingURL=index.js.map