+++ packages/adhere-ui-searchtable/lib/searchtable.d.ts
@@ -79,7 +79,7 @@ declare abstract class SearchTable extends Suspense<ISearchTableProps, ISearchTa
     /**
      * renderSearchFooterItems - 渲染SearchFooter的按钮组
      */
-    abstract renderSearchFooterItems(): Array<React.ReactElement> | null;
+    abstract renderSearchFooterItems(defaultItems: Array<React.ReactElement>): Array<React.ReactElement> | null;
     /**
      * onSearch - 进行查询
      */