+++ packages/adhere-website/src/components/ui/css/index.jsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 
 import Playground from '@/lib/Playground';
 
-import '@baifendian/adhere/lib/css.less';
+// import '@baifendian/adhere/lib/css.less';
 
 export default () => {
   const list = [];