+++ packages/adhere-ui-searchtable/changelog/CHANGELOG.md
@@ -1,7 +1,37 @@
+# 1.0.6
+
+***
+
+2021-03-21
+
+* index.less加入antd/es/pagination/style/index.less样式
+* renderSearchFooterItems加入defaultItems参数
+
+# 1.0.6
+
+***
+
+2021-03-21
+
+* **adhere-ui-searchtable**
+  - index.less加入antd/es/pagination/style/index.less样式
+  - renderSearchFooterItems加入defaultItems参数
+* **adhere-ui-searchtable**
+  - index.less加入antd/es/pagination/style/index.less样式
+  - renderSearchFooterItems加入defaultItems参数
+
+# 1.0.5
+
+***
+
+2021-03-21
+
+* 发布第一个版本
+
 # 1.0.0
 
----
+***
 
 2020-12-22
 
-- 加入代码
+* 加入代码